+++ extensions/pomodoro/src/pomodoro-menu-bar.tsx
@@ -1,5 +1,6 @@
 import { MenuBarExtra, Icon, launchCommand, LaunchType, Image, Color } from "@raycast/api";
 import { useState } from "react";
+import { FocusText, LongBreakText, ShortBreakText } from "../lib/constants";
 import {
   createInterval,
   getCurrentInterval,
@@ -33,8 +34,6 @@ export default function TogglePomodoroTimer() {
     } catch (error) {
       console.error(error);
     }
-
-    resetInterval();
   }
 
   function onStart(type: IntervalType) {
@@ -96,21 +95,21 @@ export default function TogglePomodoroTimer() {
       ) : (
         <>
           <MenuBarExtra.Item
-            title={`Focus`}
+            title={FocusText}
             subtitle={`${preferences.focusIntervalDuration}:00`}
             icon={`🎯`}
             onAction={() => onStart("focus")}
             shortcut={{ modifiers: ["cmd"], key: "f" }}
           />
           <MenuBarExtra.Item
-            title={`Short Break`}
+            title={ShortBreakText}
             subtitle={`${preferences.shortBreakIntervalDuration}:00`}
             icon={`🧘‍♂️`}
             onAction={() => onStart("short-break")}
             shortcut={{ modifiers: ["cmd"], key: "s" }}
           />
           <MenuBarExtra.Item
-            title={`Long Break`}
+            title={LongBreakText}
             subtitle={`${preferences.longBreakIntervalDuration}:00`}
             icon={`🚶`}
             onAction={() => onStart("long-break")}