+++ extensions/system-monitor/src/Power/PowerMonitor.tsx
@@ -1,4 +1,4 @@
-import { Color, List } from "@raycast/api";
+import { Icon, List } from "@raycast/api";
 import {
   getBatteryCondition,
   getBatteryLevel,
@@ -14,14 +14,15 @@ import { usePromise } from "@raycast/utils";
 
 export default function PowerMonitor() {
   const { revalidate, data: batteryLevel } = usePromise(getBatteryLevel);
+
   useInterval(revalidate, 1000);
 
   return (
     <List.Item
       id="power"
-      title={`Power`}
-      icon={{ source: "lightning.png", tintColor: Color.Yellow }}
-      accessories={[{ text: batteryLevel ? `${batteryLevel}%` : "Loading…" }]}
+      title="Power"
+      icon={Icon.Plug}
+      accessories={[{ text: batteryLevel ? `${batteryLevel} %` : "Loading…" }]}
       detail={<PowerMonitorDetail batteryLevel={batteryLevel || ""} />}
       actions={<Actions />}
     />
@@ -34,6 +35,7 @@ function PowerMonitorDetail({ batteryLevel }: { batteryLevel: string }) {
     data: isCharging,
     isLoading: isLoadingCharging,
   } = usePromise(getIsCharging);
+
   useInterval(revalidateIsCharging, 1000);
 
   const {
@@ -41,6 +43,7 @@ function PowerMonitorDetail({ batteryLevel }: { batteryLevel: string }) {
     data: batteryTime,
     isLoading: isLoadingBatteryTime,
   } = usePromise(getBatteryTime);
+
   useInterval(revalidateIsBatteryTime, 1000);
 
   const {
@@ -48,6 +51,7 @@ function PowerMonitorDetail({ batteryLevel }: { batteryLevel: string }) {
     data: timeOnBattery,
     isLoading: isLoadingTimeOnBattery,
   } = usePromise(getTimeOnBattery);
+
   useInterval(revalidateTimeOnBattery, 1000 * 60);
 
   const { data: cycleCount, isLoading: isLoadingCycleCount } = usePromise(getCycleCount);
@@ -66,7 +70,7 @@ function PowerMonitorDetail({ batteryLevel }: { batteryLevel: string }) {
       }
       metadata={
         <List.Item.Detail.Metadata>
-          <List.Item.Detail.Metadata.Label title="Battery Level" text={batteryLevel + "%"} />
+          <List.Item.Detail.Metadata.Label title="Battery Level" text={`${batteryLevel} %`} />
           <List.Item.Detail.Metadata.Label title="Charging" text={isCharging ? "Yes" : "No"} />
           <List.Item.Detail.Metadata.Label title="Cycle Count" text={cycleCount || "Loading…"} />
           <List.Item.Detail.Metadata.Label title="Condition" text={batteryCondition || "Loading…"} />