+++ extensions/homeassistant/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Home Assistant Changelog
 
+## [Update] - 2024-01-27
+
+- Add three Single Entity Menu Commands, two disabled by default
+
 ## [Update] - 2024-01-19
 
 - Make use of cache for faster loading of entities